Second pass cleanup: fix scripts, remove dead configs, update tools
Scripts:
- install-minimal.sh: remove stale .bash_it, base16-shell, .git-completion refs
- install.sh: remove dead comment
- install-dev.sh: remove unused yarn config
- update-weekly.sh: fix shebang (sh->bash), use absolute paths
- update-env.sh: fix vim path (bundle->pack), use git -C (no cd), guard virtualenvs
- rust-install.sh: exa->eza, remove deprecated racer/rustsym

Configs:
- Remove .screenrc, .abcde.conf, .Xresources, .xinitrc (unused)
- Remove TPM submodule (never loaded in tmux.conf)
- .gradle/gradle.properties: remove deprecated -XX:MaxPermSize and -Xverify:none
- .latexmkrc: cross-platform PDF previewer (Skim on mac, xdg-open on Linux)
- .wgetrc: remove stale MacPorts CA path
- fish config.fish: dynamically find latest node version
- fish env.fish: remove TERM override, guard Proton vars for Linux only

Vim plugins:
- vim-multiple-cursors -> vim-visual-multi (archived)
- w0rp/ale -> dense-analysis/ale (moved)

+++ update-env.sh
@@ -2,26 +2,20 @@
 set -x
 
 git -C ~/.oh-my-zsh pull
-#git -C ~/.bash_it pull
 
-for d in ~/.vim/bundle/*; do
-	echo "$d"
-	cd "$d" || exit
-	git fetch -p --all
-	git pull
-	cd || exit
+# Update vim native packages
+for d in ~/.vim/pack/plugins/start/*; do
+    [ -d "$d/.git" ] || continue
+    echo "$d"
+    git -C "$d" pull
 done
 
-for d in ~/.virtualenvs/*; do
-	. "$d/bin/activate"
-	pip install --upgrade pip setuptools wheel
-	pip list --outdated --format=freeze | grep -v '^\-e' | cut -d = -f 1 | xargs -n1 pip install -U
-	#pip freeze --local | grep -v '^\-e' | cut -d = -f 1  | xargs -n1 pip install -U
-	deactivate
-done
-
-#npm outdated -g --depth=0
-#npm update -g
-
-# this is now done via the deb package
-#gcloud components update -q
+# Update virtualenvs (if using virtualenvwrapper)
+if [ -d "$HOME/.virtualenvs" ]; then
+    for d in ~/.virtualenvs/*/; do
+        [ -f "$d/bin/activate" ] || continue
+        . "$d/bin/activate"
+        pip install --upgrade pip setuptools wheel
+        deactivate
+    done
+fi
